move EnableDaemon, StopDaemon, RemoveAgentArtifacts from public agent/phases to cmd/agent/internal/daemon
