--- agent/phases/reset/cleanup.go
@@ -1,53 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-package reset
-
-import (
-	"context"
-	"log/slog"
-	"path/filepath"
-
-	"github.com/Azure/unbounded/agent/phases"
-)
-
-type removeAgentArtifacts struct {
-	log *slog.Logger
-}
-
-// RemoveAgentArtifacts returns a task that removes the agent binary, install
-// script, legacy uninstall script, config directory, and temp files.
-func RemoveAgentArtifacts(log *slog.Logger) phases.Task {
-	return &removeAgentArtifacts{log: log}
-}
-
-func (t *removeAgentArtifacts) Name() string { return "remove-agent-artifacts" }
-
-func (t *removeAgentArtifacts) Do(_ context.Context) error {
-	t.log.Info("removing agent binaries and configuration")
-
-	// Remove known file paths.
-	for _, path := range []string{
-		"/usr/local/bin/unbounded-agent",
-		"/usr/local/bin/unbounded-agent-install.sh",
-		"/usr/local/bin/unbounded-agent-uninstall.sh",
-	} {
-		removeFileIfExists(t.log, path)
-	}
-
-	// Remove directories.
-	for _, dir := range []string{
-		"/etc/unbounded/agent",
-		"/tmp/unbounded-agent",
-	} {
-		removeAllIfExists(t.log, dir)
-	}
-
-	// Remove temp config files matching /tmp/unbounded-agent-config.*.json.
-	matches, _ := filepath.Glob("/tmp/unbounded-agent-config.*.json") //nolint:errcheck // Pattern is valid; only errors on malformed globs.
-	for _, m := range matches {
-		removeFileIfExists(t.log, m)
-	}
-
-	return nil
-}
