+++ agent/phases/reset/nspawn.go
@@ -37,35 +37,11 @@ func (t *removeNSpawnConfig) Do(_ context.Context) error {
 	return nil
 }
 
-type removeAppliedConfig struct {
-	log         *slog.Logger
-	machineName string
-}
-
-// RemoveAppliedConfig returns a task that removes the applied agent config
-// file for the named machine. Errors are logged but do not fail the task.
-func RemoveAppliedConfig(log *slog.Logger, machineName string) phases.Task {
-	return &removeAppliedConfig{log: log, machineName: machineName}
-}
-
-func (t *removeAppliedConfig) Name() string { return "remove-applied-config" }
-
-func (t *removeAppliedConfig) Do(_ context.Context) error {
-	path := goalstates.AppliedConfigPath(t.machineName)
-	removeFileIfExists(t.log, path)
-
-	checksumPath := goalstates.AppliedConfigChecksumPath(t.machineName)
-	removeFileIfExists(t.log, checksumPath)
-
-	return nil
-}
-
 // CleanupMachine returns a composite task that removes all artifacts of an
-// nspawn machine: its nspawn configuration, rootfs, and applied config.
+// nspawn machine: its nspawn configuration and rootfs.
 func CleanupMachine(log *slog.Logger, machineName string) phases.Task {
 	return phases.Serial(log,
 		RemoveNSpawnConfig(log, machineName),
 		RemoveMachine(log, machineName),
-		RemoveAppliedConfig(log, machineName),
 	)
 }
